refactor(podman): move podman-install.ts to proper folder
